Break up request adjudicator tests to separate file for readability
